Fix: let f ?(f = fun _ -> 1) () = f
The piece of code above was incorrectly colorized.
Before this commit, the second parenthesis was capture by the fun
so the optional argument assignement was never ended.

There was a parenthesis "imbalance" in the way anonymous functions
were captured.

I hope that the look-ahead used does not make the syntax coloring
too slow.